fix(backend): use typeshare-safe integers
Prevent typeshare-core from rejecting 64-bit integers by using i32 for pagination and expiry fields.

+++ backend/src/models/pagination.rs
@@ -4,21 +4,21 @@ use typeshare::typeshare;
 #[typeshare]
 #[derive(Debug, Deserialize)]
 pub struct PaginationQuery {
-    pub limit: Option<i64>,
-    pub offset: Option<i64>,
+    pub limit: Option<i32>,
+    pub offset: Option<i32>,
 }
 
 #[typeshare]
 #[derive(Debug, Serialize)]
 pub struct PaginatedResponse<T> {
     pub data: Vec<T>,
-    pub total: i64,
-    pub limit: i64,
-    pub offset: i64,
+    pub total: i32,
+    pub limit: i32,
+    pub offset: i32,
 }
 
 impl<T> PaginatedResponse<T> {
-    pub fn new(data: Vec<T>, total: i64, limit: i64, offset: i64) -> Self {
+    pub fn new(data: Vec<T>, total: i32, limit: i32, offset: i32) -> Self {
         Self {
             data,
             total,
